+++ tests/Unit/Property/KeyframeSelectorTest.php
@@ -6,6 +6,7 @@ namespace Sabberworm\CSS\Tests\Unit\Property;
 
 use PHPUnit\Framework\TestCase;
 use Sabberworm\CSS\Property\KeyframeSelector;
+use Sabberworm\CSS\Property\Selector\CompoundSelector;
 
 /**
  * @covers \Sabberworm\CSS\Property\KeyframeSelector
@@ -23,5 +24,17 @@ final class KeyframeSelectorTest extends TestCase
         $result = $subject->getArrayRepresentation();
 
         self::assertSame('KeyframeSelector', $result['class']);
+    }
+
+    /**
+     * @test
+     */
+    public function getArrayRepresentationIncludesComponent(): void
+    {
+        $subject = new KeyframeSelector([new CompoundSelector('50%')]);
+
+        $result = $subject->getArrayRepresentation();
+
+        self::assertSame('50%', $result['components'][0]['value']);
     }
 }